# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PacienteRequestDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"dataNascimento": {
						"type": "string"
					},
					"telefone": {
						"type": "number",
						"format": "double"
					},
					"endereco": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"cpf",
					"dataNascimento",
					"telefone",
					"endereco"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PacienteDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"dataNascimento": {
						"type": "string"
					},
					"telefone": {
						"type": "number",
						"format": "double"
					},
					"endereco": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"cpf",
					"dataNascimento",
					"telefone",
					"endereco"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MedicoRequestDto": {
				"properties": {
					"crm": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"telefone": {
						"type": "number",
						"format": "double"
					},
					"endereco": {
						"type": "string"
					}
				},
				"required": [
					"crm",
					"name",
					"telefone",
					"endereco"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MedicoDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"crm": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"telefone": {
						"type": "number",
						"format": "double"
					},
					"endereco": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"crm",
					"name",
					"telefone",
					"endereco"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExameRequestDto": {
				"properties": {
					"pacienteCpf": {
						"type": "number",
						"format": "double"
					},
					"medicoID": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"type": "string"
					}
				},
				"required": [
					"pacienteCpf",
					"medicoID",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExameDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"pacienteCpf": {
						"type": "number",
						"format": "double"
					},
					"medicoID": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"pacienteCpf",
					"medicoID",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
		"/paciente": {
			"post": {
				"operationId": "CadastrarPaciente",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Paciente"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PacienteRequestDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizaPaciente",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Paciente"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PacienteDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletaPaciente",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Paciente"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PacienteDto"
							}
						}
					}
				}
			}
		},
		"/paciente/busca/{id}": {
			"get": {
				"operationId": "GetPaciente",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Paciente"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/paciente/todos": {
			"get": {
				"operationId": "GetPacientes",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Paciente"
				],
				"security": [],
				"parameters": []
			}
		},
		"/medico": {
			"post": {
				"operationId": "CadastrarMedico",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Medico"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MedicoRequestDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizaMedico",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Medico"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MedicoDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletaMedico",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Medico"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MedicoDto"
							}
						}
					}
				}
			}
		},
		"/medico/busca/{crm}": {
			"get": {
				"operationId": "GetMedico",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Medico"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "crm",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/medico/todos": {
			"get": {
				"operationId": "GetMedicos",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Medico"
				],
				"security": [],
				"parameters": []
			}
		},
		"/exame": {
			"post": {
				"operationId": "CadastrarExame",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Exame"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExameRequestDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizaExame",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Exame"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExameDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletaExame",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Exame"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ExameDto"
							}
						}
					}
				}
			}
		},
		"/exame/busca/{pacienteID}": {
			"get": {
				"operationId": "GetExame",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Exame"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "pacienteID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/exame/todos": {
			"get": {
				"operationId": "GetExames",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Exame"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3040/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}